vfs: add two sub-type Vfs 'UrlVfs' & 'LocalVfs'
vfs: refactor Vfs store/load key-values into own KeyValueStorage

context: add new urlVfs & applicationVfs types
